@@ -393,7 +393,7 @@ def compile(  # noqa A001
     pattern: AnyStr | Pattern[AnyStr] | Bre[AnyStr],
     flags: int = 0,
     auto_compile: bool | None = None
-) -> 'Bre[AnyStr]':
+) -> Bre[AnyStr]:
     """Compile both the search or search and replace into one object."""
 
     if isinstance(pattern, Bre):
